+++ CMakeLists.txt
@@ -140,6 +140,7 @@ set(libsrc
     src/plugins_types_decimal64.c
     src/plugins_types_empty.c
     src/plugins_types_enumeration.c
+    src/plugins_types_identityref.c
     src/plugins_types_integer.c
     src/plugins_types_string.c
     src/plugins_exts.c
